service/requirements.txt
Flask==2.2.3
py-healthcheck==1.10.1